requirements.txt
openai>=1.30.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
joblib>=1.3.0
matplotlib>=3.8.0
jsonschema>=4.21.0
python-dotenv>=1.0.1
pytest>=8.0.0
tabulate>=0.9.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
